refactor(openai-compatible): extract runIteration concerns into query/ modules
Extract tool dispatch and message append handling from query.ts into
query/dispatch-append.ts without changing behavior.
